requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-autoslug==1.9.8
django-cities-light==3.9.2
django-countries==7.5.1
django-summernote==0.8.20.0
gunicorn==20.1.0
progressbar2==4.2.0
psycopg2==2.9.5
python-utils==3.5.2
pytz==2022.7.1
sqlparse==0.4.3
Unidecode==1.3.6
